# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /trade/order:
    post:
      operationId: placeTradeOrder
      tags: [trade]
      summary: Execute a market order immediately at the current live price
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlaceTradeOrderInput" }
      responses:
        "201": { description: Order filled, content: { application/json: { schema: { type: object, properties: { order: { $ref: "#/components/schemas/TradeOrder" }, balance: { type: number } } } } } }
...
  /trade/pending-order:
    post:
      operationId: placePendingOrder
      tags: [trade]
      summary: Place a resting limit or stop order (filled by a 5s background poller once triggered)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlacePendingOrderInput" }
      responses:
        "201": { description: Order placed, content: { application/json: { schema: { $ref: "#/components/schemas/PendingOrder" } } } }
  /trade/pending-order/{id}/cancel:
    post:
      operationId: cancelPendingOrder
      tags: [trade]
      summary: Cancel an open pending order
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Cancelled, content: { application/json: { schema: { $ref: "#/components/schemas/PendingOrder" } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token returned by POST /auth/verify-otp.
  schemas:
...
    TradeOrder:
      type: object
      properties:
        id: { type: string }
        market: { type: string, enum: [crypto, stock] }
        symbol: { type: string }
        side: { type: string, enum: [BUY, SELL] }
        qty: { type: number }
        price: { type: number }
        value: { type: number }
        fee: { type: number }
        createdAt: { type: string }
        paper: { type: boolean }
...
    PendingOrder:
      type: object
      properties:
        id: { type: string }
        market: { type: string, enum: [crypto, stock] }
        symbol: { type: string }
        side: { type: string, enum: [buy, sell] }
        kind: { type: string, enum: [limit, stop] }
        quantity: { type: number }
        triggerPrice: { type: number }
        status: { type: string, enum: [open, filled, cancelled] }
        note: { type: string }
        filledPrice: { type: number }
        createdAt: { type: string }
        filledAt: { type: ["string", "null"] }
...
    PlaceTradeOrderInput:
      type: object
      required: [market, symbol, side, quantity]
      properties: { market: { type: string, enum: [crypto, stock] }, symbol: { type: string }, side: { type: string, enum: [BUY, SELL] }, quantity: { type: string } }
    PlacePendingOrderInput:
      type: object
      required: [market, symbol, side, kind, quantity, triggerPrice]
      properties: { market: { type: string, enum: [crypto, stock] }, symbol: { type: string }, side: { type: string, enum: [BUY, SELL] }, kind: { type: string, enum: [LIMIT, STOP] }, quantity: { type: number }, triggerPrice: { type: number } }
...